Move about component to pages folder
